refactor(cli-core): move handle_run + handle_test to leo_cli_core::commands
`handle_run` and `handle_test` (with `discover_test_functions`) move
from `crates/leo/src/cli/commands/{run,test}.rs` to
`crates/leo-cli-core/src/commands/{run,test}.rs`. The CLI's `LeoRun`
and `LeoTest` clap structs are now thin wrappers — they collect flags,
trigger `LeoBuild` for prelude, then call into core.

Supporting moves (also out of `crates/leo` into `leo-cli-core`):
- `parse_input` + `validate_cli_literal` + helpers (from `commands/mod.rs`).
- `print_program_source`, `check_edition_constructor_requirements`,
  `load_extra_programs_into_vm` (from `common/util.rs`).
- `load_latest_programs_from_network` (from `common/query.rs`); the CLI
  copy is now a thin wrapper that converts `&Context` → `&Path`.
- `get_endpoint`, `get_network`, `get_private_key`, `get_is_devnet`
  (from `common/options.rs`).
- `RunOutput`, `TestOutput`, `TestResult` (from `common/output.rs`) —
  `LeoRun::Output = RunOutput` / `LeoTest::Output = TestOutput` paths
  keep working via re-export.

`handle_run` takes a `RunArgs<'_>` value (rather than `&LeoRun`) so it
doesn't depend on the clap struct. `handle_test` takes
`(Package, &str, NetworkName, bool)`.

`leo-cli-core` Cargo.toml gains `aleo-std-storage`, `colored`, `rand`
as native-only deps.
